requirements.txt
streamlit>=1.31
python-dotenv>=1.0
passlib>=1.7.4
openai>=1.40.0
pytest>=8.0
